doc: driver-api: add loop file format subsystem API documentation
The entire API of the file format subsystem for loop devices is documented
in the reST kernel documentation format. The documentation deals with the
description of the internal API of the file format subsystem to access its
functionality and adds a section on how to write own loop file format
drivers using the driver API of the subsystem.

Signed-off-by: Manuel Bentele <development@manuel-bentele.de>