block: loop: add QCOW2 loop file format driver (read-only)
The QCOW2 file format is added as a new file format driver module to the
existing loop device file format subsystem. The implementation of the QCOW2
file format is based on the original implementation from the QEMU project
and was ported to the Linux kernel space.

The current implementation of the QCOW2 file format supports the reading of
normal QCOW2 disk images as well as the reading of sparsed or compressed
QCOW2 images. Write support is missing and is not ported yet. Discard,
flush and reading or writing aio is missing, too, and can be implemented
together with QCOW version 1 support in the future.

Signed-off-by: Manuel Bentele <development@manuel-bentele.de>